zonalStatistics: move pdf logic to separate file
